Part2
Find the first non-covered item within a bounding box.

It's...big. 4Mx4M so even a naive "we'll just iterate 4M rows" is way too slow. 2/sec on a single core so even getting to 24 cores would still take a day.

Realized all we care is the interval covered on a line. Once we find the first space not covered we're good.

Wrote a sliding interval detector and gets done in <12s.