day16 part2
This isn't right but...this code happens to work for the real answer, just not the test data.

Thought finding the best P(6,3) of the original and then P(3,3) of the diff would do it.

On test data this doesn't optimize since there are combos that are bigger than "biggest + biggest remainder".