day21 part2
Need to figure out what value to make humn so it matches the other subtree from root.

Do this by adding an operation state Unknown and setting humn to this. Then refuse to compute operations we don't know so we end up with one branch from root which isn't computed.

Find our match value as the subtree from root which has a value.

Then recurse through the other tree inverting each op against the current value.

i.e. initial value is 150 and first op is "divide by 4" which means "multiple by 4" to invert it.

Then go through each layer doing this until we run out of nodes and try to process humn. At that point we know the algebra answer for this.