day22 part2
Wow...A cube..Fun.

Implemented by arts/crafts the cube in graph paper.

Then we hard code all the face transitions and turn that into a function. Thankfully with a different translation set we can do part1 with the same function.

I had 2 sets of problems:

1. Translations for N->S/S-N and E-W/W-E didn't take into account you flip. So one axis is the mirror of it's original. That made my answer for part2 continually wrong.

Additionally somehow I wrote the wrong translations for the same input but when combined with the above error actually gave the right answer....probably because it's 4x4 blocks not a lot of walls so you can end up in the right spot by accident.

To debug I finally had to download someone elses solution and print out their walk to see my translation errors.

TODO: Write something which can fold the input into a cube and figure out the translations dynamically.